backend/requirements.txt
Flask
flask_cors
paho-mqtt
gunicorn
